Visual Editor: Extracted toolbar code to VisualEditorToolbar
1 Extracted toolbar code to VisualEditorToolbar
2 Dropped OrderedList and UnorderedList feature from editor(buggy and AnkiDesktop doesn't have this feature)
3 Dropped RegisterMediaForWebView and used MediaRegistration class
4 audio fix
