+++ AnkiDroid/src/main/java/com/ichi2/anki/multimediacard/visualeditor/SummerNoteVisualEditor.java
@@ -1,3 +1,4 @@
+//noinspection MissingCopyrightHeader #5288
 /*
 MIT License
