+++ AnkiDroid/src/main/res/values/16-multimedia-editor.xml
@@ -52,7 +52,6 @@
     <string name="multimedia_editor_field_editing_image">Image</string>
     <string name="multimedia_editor_field_editing_audio">Audio recording</string>
     <string name="multimedia_editor_field_editing_audio_clip">Audio Clip</string>
-    <string name="multimedia_editor_field_editing_visual_editor">Visual Editor</string>
     <string name="multimedia_editor_field_editing_done" comment="Label of a button to tell the app that the current action is 'Done'.">Done</string>
 
     <!-- Text field editing -->
@@ -124,27 +123,26 @@
     <string name="title_activity_visual_editor">Visual Editor</string>
 
     <!-- Visual Editor -->
-    <string name="visual_editor_tooltip_bold">Bold</string>
-    <string name="visual_editor_tooltip_italic">Italic</string>
-    <string name="visual_editor_tooltip_underline">Underline</string>
-    <string name="visual_editor_tooltip_cloze">Cloze Deletion</string>
-    <string name="visual_editor_tooltip_clear_formatting">Clear Formatting</string>
-    <string name="visual_editor_tooltip_bullet_list">Bulleted List</string>
-    <string name="visual_editor_tooltip_numbered_list">Numbered List</string>
+    <string name="visual_editor_tooltip_bold">Text Bold</string>
+    <string name="visual_editor_tooltip_italic">Text Italic</string>
+    <string name="visual_editor_tooltip_underline">Text Underline</string>
+    <string name="visual_editor_tooltip_cloze">Deletion Cloze</string>
+    <string name="visual_editor_tooltip_clear_formatting">Formatting Clear</string>
     <string name="visual_editor_tooltip_text_color">Text Color</string>
     <string name="visual_editor_tooltip_background_color">Background Color</string>
     <string name="visual_editor_tooltip_horizontal_line">Horizontal Line</string>
-    <string name="visual_editor_tooltip_align_left">Left Align</string>
-    <string name="visual_editor_tooltip_align_center">Center Align</string>
-    <string name="visual_editor_tooltip_align_right">Right Align</string>
+    <string name="visual_editor_tooltip_align_left">Left Alignment</string>
+    <string name="visual_editor_tooltip_align_center">Center Alignment</string>
+    <string name="visual_editor_tooltip_align_right">Right Alignment</string>
     <string name="visual_editor_tooltip_align_justify">Justify</string>
     <string name="visual_editor_tooltip_view_source">View Source</string>
-    <string name="visual_editor_tooltip_add_image">Add Image</string>
+    <string name="visual_editor_tooltip_add_image">Image Add</string>
     <string name="visual_editor_tooltip_cut_image">Cut</string>
-    <string name="visual_editor_tooltip_delete_image">Delete Image</string>
-    <string name="visual_editor_tooltip_record_audio">Record Audio</string>
-    <string name="visual_editor_tooltip_mathjax">Insert MathJax Equation</string>
+    <string name="visual_editor_tooltip_delete_image">Image Delete</string>
+    <string name="visual_editor_tooltip_record_audio">Audio Record</string>
+    <string name="visual_editor_tooltip_mathjax">MathJax Equation Add</string>
     <string name="visual_editor_white_board">Open WhiteBoard</string>
-    <string name="visual_editor_add_custom_buttons">Add Custom buttons</string>
+    <string name="visual_editor_custom_buttons">Custom Buttons</string>
+    <string name="visual_editor_failed">Unable to start visual editor</string>
 
 </resources>
